Rather than clipping tooltip, snap to sides
This commit introduces a ref to a wrapper div that I'm using so that I
can call `getBoundingClientRect` and do some math. The results of that
math help me know if the tip is going to end up outside the window and I
can adjust it's offset from the anchor accordingly.