[FIX Link] Remove the purple border from <Link>
While the border is helpful when focusing on a link and using the tab
key to move forward, it looks too disruptive for most of our default
links. Let's just remove it for now and re-visit when we think we
really need it.

Also adding href to <Link> otherwise browsers do not apply default
styles to link tags.