Move x-axis-label to its own container
to solve the vertical align problem. It is a hybrid of using flexbox to
create a container for those labels and absolute positioning to center
align the labels. The only hack was to set min-height for the label
container since its content is absolutely positioned and doesn't take
space. I think it is fine for our use case because in our design system
spec the labels are <Sans size="2"/> so setting min-height to space(2)
solves the problem but it is possible to pass a component as label which
in that case this solution doesn't work