Add ts type-checking task for iOS (see commit desc or #491 for details)
See this comment for the initial inspiration: https://github.com/microsoft/TypeScript/issues/8328#issuecomment-219583889

This change adds a new ts config file that overwrites it's module resolution behavior. When importing a relative file
it first tries to find a *.ios version of the file. This is in line with how react-native handles platform specific implementations.