Update Component Spec template to include web Zeplin link
This commit updates the Component Step template to prompt for both
Zeplin links: web and app. Those interested in the spec are not always
on a system with the Zeplin app installed. This prompts issue creators
to provide both links so that the spec can be accessed by a wider range
of viewers and devices.