WIP: next next/prev buttons working
This diff is me tinkering with getting the next/prev buttons working
with the getHref function. I was getting linking errors and just
couldn't quite get it to work the way I expected it to.

One thing I _think_ I discovered is that while we provide a previous
property in the page cursors interface, we do not for next and so I
wrote a little ditty to compute this and I don't know if it's correct.