fix: be defensive when animating Collapse async
I noticed a flaky error happens on CI sometimes when running tests:

```
TypeError: Cannot read properties of null (reading 'style')

  at style (node_modules/@artsy/palette/src/elements/Collapse/Collapse.tsx:56:26)
```

When running tests individually, this error doesn't happen. My theory is
that the callback in `setTimeout` is not finished in the same test and
instead is leaked into the subsequent test and fails with the `null`
error. Perhaps the element is already gone by then.

I'd like to investigate more on improving the tests, and this commit
makes the callback more defensive.

https://app.circleci.com/pipelines/github/artsy/force/52795/workflows/8bb542a0-ed21-4864-bd46-053663728635/jobs/395284