Reproduce the Autocomplete bug within Drawer
[Slack
thread](https://artsy.slack.com/archives/C05F8TNKGAV/p1748537455230889)

When using the autocomplete input within a drawer positioned to the
right, we're bumping into issues with the position of the options. It
doesn't always display right below the component (sometimes it doesn't
even display on any visibile spot).

As pointed out by Ole, this seems to be a conflict between the
Autocomplete options relying on `position: fixed` and the drawer relying
on the `transform` property, leading to an uncanny relationship (more
about this
[here](https://dev.to/salilnaik/the-uncanny-relationship-between-position-fixed-and-transform-property-32f6)).